Спринт 10. Добавила Glide.
