+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker
 
 
 // Класс для хранения информации об одном треке
-class Track (
+data class Track (
     var trackName: String, // Название композиции
     var artistName: String, // Имя исполнителя
     var trackTime: String, // Продолжительность трека
