+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -2,10 +2,12 @@ package com.example.playlistmaker
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import com.bumptech.glide.Glide
 
 class TrackAdapter : RecyclerView.Adapter<TrackViewHolder>() {
+   // private val imageUrl = "https://img.freepik.com/free-vector/open-blue-book-white_1308-69339.jpg"
     var items: MutableList<Track> = mutableListOf()
         set(value) {
             field = value.toMutableList()
@@ -14,6 +16,16 @@ class TrackAdapter : RecyclerView.Adapter<TrackViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
         val itemView = LayoutInflater.from(parent.context).inflate(R.layout.track_layout, parent, false)
+
+       /* val image = parent.findViewById<ImageView>(R.id.image)
+
+        Glide.with(itemView).load(imageUrl).into(image)
+
+        Glide.with(itemView)
+            .load(imageUrl)
+            .placeholder(R.drawable.place_holder)
+            .into(image)
+*/
         return TrackViewHolder(itemView)
     }
