+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
 
     <TextView
         style="@style/TextTitle"
-        android:textSize="@dimen/text_size"
+        android:textSize="@dimen/text_size_22"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/app_name"
