+++ app/src/main/res/layout/track_layout.xml
@@ -1,56 +1,65 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:orientation="horizontal"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content">
 
-    <TextView
-        style="@style/TextTitle"
-        android:id="@+id/tvTrackName"
-        android:textSize="@dimen/text_size"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="title"
-        android:textColor="@color/white"
-        android:layout_gravity="left"
-        android:layout_marginLeft="@dimen/margin_8"
-        android:layout_marginTop="@dimen/margin_6"
-        android:layout_marginBottom="@dimen/margin_4"/>
-    <TextView
-        style="@style/TextTitle"
-        android:id="@+id/tvArtistName"
-        android:textSize="@dimen/text_size"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="artist"
-        android:textColor="@color/white"
-        android:layout_gravity="left"
-        android:layout_marginLeft="@dimen/margin_8"
-        android:layout_marginTop="@dimen/margin_6"
-        android:layout_marginBottom="@dimen/margin_4"/>
-    <TextView
-        style="@style/TextTitle"
-        android:id="@+id/trackTime"
-        android:textSize="@dimen/text_size"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="time"
-        android:textColor="@color/white"
-        android:layout_gravity="left"
-        android:layout_marginLeft="@dimen/margin_8"
-        android:layout_marginTop="@dimen/margin_6"
-        android:layout_marginBottom="@dimen/margin_4"/>
-    <TextView
-        style="@style/TextTitle"
-        android:id="@+id/artworkUrl100"
-        android:textSize="@dimen/text_size"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="url"
-        android:textColor="@color/white"
-        android:layout_gravity="left"
-        android:layout_marginLeft="@dimen/margin_8"
-        android:layout_marginTop="@dimen/margin_6"
-        android:layout_marginBottom="@dimen/margin_4"/>
-
+    <ImageView
+        android:id="@+id/image"
+        android:layout_width="@dimen/margin_45"
+        android:layout_height="@dimen/margin_45"
+        tools:src="@drawable/support"
+        android:layout_marginVertical="@dimen/margin_8"
+        android:layout_marginStart="@dimen/margin_7"
+        android:layout_marginEnd="@dimen/margin_8"
+        android:layout_gravity="center">
+    </ImageView>
+    <!--tools:src="@drawable/support"-->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:layout_gravity="center_vertical">
+        <TextView
+            android:id="@+id/tvTrackName"
+            style="@style/TextTrackName"
+            android:textSize="@dimen/text_size_16"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="start|center_vertical"
+            android:layout_marginTop="@dimen/margin_14"
+            android:layout_marginBottom="@dimen/margin_1"
+            tools:text="title title title title"
+            android:textColor="@color/edit_text_track_big"
+            />
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content">
+            <TextView
+                style="@style/TextArtistName"
+                android:id="@+id/tvArtistName"
+                android:textSize="@dimen/text_size_11"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="start|center_vertical"
+                android:text="artist firstName LastName"
+                android:textColor="@color/edit_text_track_small"/>
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:src="@drawable/ic">
+            </ImageView>
+            <TextView
+                style="@style/TextArtistName"
+                android:id="@+id/tvTrackTime"
+                android:textSize="@dimen/text_size_11"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="start|center_vertical"
+                tools:text="4:06"
+                android:textColor="@color/edit_text_track_small"/>
+               </LinearLayout>
+    </LinearLayout>
 </LinearLayout>
