+++ app/src/main/res/values-night/colors.xml
@@ -17,6 +17,9 @@
     <color name="edit_background">#FFFFFF</color>
     <color name="edit_textHint">#1A1B22</color>
     <color name="icon_for_edit">#1A1B22</color>
+
+    <color name="edit_text_track_big">#FFFFFF</color>
+    <color name="edit_text_track_small">#FFFFFF</color>
 </resources>
 <!-- <color name="selector_checked">#3772E7</color>
     <color name="selector_nonchecked">#AEAFB4</color> -->
