+++ app/src/main/res/values/colors.xml
@@ -19,6 +19,9 @@
     <color name="icon_for_edit">#AEAFB4</color>
 
     <color name="cursor_edittext">#3772E7</color>
+
+    <color name="edit_text_track_big">#000000</color>
+    <color name="edit_text_track_small">#AEAFB4</color>
 </resources>
 
 <!--     <color name="selector_checked">#3772E7</color>
