+++ app/src/main/res/values/dimens.xml
@@ -2,12 +2,15 @@
 <resources>
     <dimen name="small_padding">4dp</dimen>
     <dimen name="medium_padding">8dp</dimen>
-    <dimen name="text_size">22sp</dimen>
-    <dimen name="text_size_menu_settings">16sp</dimen>
+    <dimen name="text_size_11">11sp</dimen>
+    <dimen name="text_size_16">16sp</dimen>
+    <dimen name="text_size_22">22sp</dimen>
     <dimen name="corner_radius_16">16dp</dimen>
     <dimen name="corner_radius_8">8dp</dimen>
+    <dimen name="margin_1">1dp</dimen>
     <dimen name="margin_4">4dp</dimen>
     <dimen name="margin_6">6dp</dimen>
+    <dimen name="margin_7">7dp</dimen>
     <dimen name="margin_8">8dp</dimen>
     <dimen name="margin_10">10dp</dimen>
     <dimen name="margin_12">12dp</dimen>
@@ -17,6 +20,7 @@
     <dimen name="margin_21">21dp</dimen>
     <dimen name="margin_28">28dp</dimen>
     <dimen name="margin_32">32dp</dimen>
+    <dimen name="margin_45">45dp</dimen>
     <dimen name="margin_48">48dp</dimen>
     <dimen name="margin_56">56dp</dimen>
     <dimen name="button_size">24dp</dimen>
