+++ app/src/main/res/values/styles.xml
@@ -5,7 +5,7 @@
         <item name="android:layout_height">0dp</item>
 
         <item name="android:textColor">@color/text_button_main_window</item>
-        <item name="android:textSize">@dimen/text_size</item>
+        <item name="android:textSize">@dimen/text_size_22</item>
         <item name="android:textAllCaps">false</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
 
@@ -20,9 +20,21 @@
         <item name="iconTint">@color/text_button_main_window</item>
     </style>
     <style name="TextTitle">
-        <item name="android:textSize">@dimen/text_size</item>
+        <item name="android:textSize">@dimen/text_size_22</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="android:textFontWeight">500</item>
+    </style>
+    <style name="TextTrackName">
+        <item name="android:textSize">@dimen/text_size_16</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="lineHeight">18.75sp</item>
+    </style>
+    <style name="TextArtistName">
+        <item name="android:textSize">@dimen/text_size_11</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="lineHeight">12.89sp</item>
     </style>
     <style name="FrameLayoutSettingsWindow">
         <item name="android:layout_width">match_parent</item>
@@ -42,7 +54,7 @@
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/text_size_menu_settings</item>
+        <item name="android:textSize">@dimen/text_size_16</item>
         <item name="android:textAllCaps">false</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
         <item name="android:layout_gravity">center_vertical</item>
@@ -68,6 +80,7 @@
         <item name="android:paddingRight">@dimen/margin_10</item>-->
         <item name="android:drawablePadding">@dimen/margin_8</item>
         <item name="android:layout_gravity">center_vertical</item>
+        <item name="lineHeight">18.75sp</item>
     </style>
 </resources>
